Fixes some missing things

+++ src/desktop/MockData/SPIAccelerometerDataInternal.h
@@ -12,8 +12,8 @@ class SPIAccelerometerData : DataBase {
   HAL_Bool GetActive();
   void SetActive(HAL_Bool active);
 
-  uint8_t GetRange();
-  void SetRange(uint8_t range);
+  uint32_t GetRange();
+  void SetRange(uint32_t range);
 
   double GetX();
   void SetX(double x);
@@ -27,7 +27,7 @@ class SPIAccelerometerData : DataBase {
   virtual void ResetData() override;
  private:
   std::atomic<HAL_Bool> m_active = false;
-  std::atomic<uint8_t> m_range = 0;
+  std::atomic<uint32_t> m_range = 0;
   std::atomic<double> m_x = 0.0;
   std::atomic<double> m_y = 0.0;
   std::atomic<double> m_z = 0.0;
